Data and time formats added to the class configuration.
Some column variables of the Spec2 corrected.

Some error handlers added to the class SampleData.

References to some self.properties from SampleData class substitude by
new variables. Was it a good idea?

equipment variable added to WriteTCT procedure

Procedure CombineDateTime created to handle the date and time of analysis
of each sample. It returns an datetime object. A example of use of this
procudedure is to combine the date of analysis and the time of the
first cycle.