A standard output extension ("dynamic.exp") was added and is used by the WriteTXT procedure. This is the extension created by Evaluation, the program used to export data files created by the Neptune. This avoid additional modifications in Chronus code.
The date of analysis and the time of each cycle in the files split will be updated to the format supported by Chronus. Something similar can be a future improvement in this python code: files produced by other equipment can be modified so Chronus will be able to handle them (the alternative is to create new date formats in Chronus VBA code.

A different way of checking if the files were already split was implemented. The name of the new file is compared to those in files_split folder. If any of these file has the exact name, so the analysis files were already split.

A bug in the CombineDateTime procedure was fixed. Now it returns an object with the correct date and time.