MainUPb
This module was the most modified. The objective was to allow Chronus to
handle data from laboratories that records blank and sample analyses in
the same files. Many public variables were created, most of them related
to the new data required in the Box1_start and Box4_Addresses userforms.
Examples:
 - Isotope202Analyzed_UPb - Isotope analyzed?
 - BlanksRecordedSamples_UPb - Samples and blanks recorded in the same
 file?
 - Detector202_UPb - Type of detector used (MIC or Faraday)
---------------------------

Function RawRanges and UpdateRawRanges - These are two of the most
important procedures created in this update. They take the range
addresses provided by the users and update them based on the number of
cycles of each sample. This important because if the blanks were
recorded with the samples, the number of cycles can be different in each
analysis.
----------------------

PublicVariables_MainSh - Checks if the main sheets created by Chronus
are present in the workbook being processed.